add: create function the create, put and delete for clientes

+++ ativ1_prisma/src/controllers/peopleController.ts
@@ -1,52 +1,88 @@
 import { PrismaClient } from "@prisma/client";
+import { error } from "console";
 import { Request, Response } from "express";
 
 const prisma = new PrismaClient();
 
-export const getPeople = async(req: Request, res: Response) => {
-    const peoples = await prisma.pessoa.findMany()
+export const getPeople = async (req: Request, res: Response) => {
+    const peoples = await prisma.pessoa.findMany({
+        include: {
+            telefone: true
+        }
+    })
     // debug
     console.log("Pessoas Encontradas:", peoples)
     res.json(peoples)
 }
 
-// export const getOnePeople = async(req: Request, res: Response) => {
-//     const { id } = req.params
-//     console.log(id)
-//     const people = await prisma.pessoa.findUnique({
-//         where: {
-//             id: Number(id)
-//         }
-//     })
-//     // debug
-//     console.log("Pessoas Encontradas:", people)
-//     res.json(people)
-// }
-
-export const createPeople = async(req: Request, res: Response) => {
-    const { nome, email, telefone } = req.body
-    const people = await prisma.pessoa.create({
-        data: {nome, email},
-    })
-    await prisma.telefone.create({
-        data: {
-            numero: telefone,
-            pessoaId: people.id
+export const getOnePeople = async (req: Request, res: Response) => {
+    const { id } = req.params
+    console.log(id)
+    const people = await prisma.pessoa.findUnique({
+        where: {
+            id: Number(id)
         }
     })
-    const peopleWithTelefone = await prisma.pessoa.findUnique({
-        where: {id: people.id},
-        include: {
-            telefone: true
-        }
-    })
-    res.status(201).json(peopleWithTelefone)
+    // debug
+    console.log("Pessoas Encontradas:", people)
+    res.json(people)
 }
 
-export const updatePeople = async(req: Request, res: Response) => {
-    const { id } = req.params
-    const { nome, email} = req.body
+export const createPeople = async (req: Request, res: Response) => {
+    try {
+        const { nome, email, telefone } = req.body
+        if (!nome || !email || !telefone) {
+            return res.status(400).json({
+                error: "Nome, email e telefone são obrigatórios!"
+            });
+        }
+        const people = await prisma.pessoa.create({
+            data: { nome, email },
+        })
+        await prisma.telefone.create({
+            data: {
+                numero: telefone,
+                pessoaId: people.id
+            }
+        })
+        const peopleWithTelefone = await prisma.pessoa.findUnique({
+            where: { id: people.id },
+            include: {
+                telefone: true
+            }
+        })
+        res.status(201).json(peopleWithTelefone)
+    } catch (e) {
+        console.error("Erro ao criar Cliente:", e);
+    }
+}
 
+export const updatePeople = async (req: Request, res: Response) => {
+    const { id } = req.params
+    const { nome, email, telefone } = req.body
+    console.log(nome, email, telefone)
+
+    const existeTelfone = await prisma.telefone.findUnique({
+        where: {
+            pessoaId: Number(id)
+        }
+    })
+    console.log(existeTelfone)
+    if (!existeTelfone) {
+        await prisma.telefone.create({
+            data: {
+                numero: telefone,
+                pessoaId: Number(id)
+            }
+        })
+    } else {
+        await prisma.telefone.update({
+            where: { pessoaId: Number(id) },
+            data: {
+                numero: telefone,
+            }
+        })
+    }
     const newPeople = await prisma.pessoa.update({
         where: { id: Number(id) },
         data: {
@@ -54,13 +90,38 @@ export const updatePeople = async(req: Request, res: Response) => {
             email
         }
     })
-    res.json(newPeople)
+
+    const peopleWithTelefone = await prisma.pessoa.findUnique({
+        where: { id: newPeople.id },
+        include: {
+            telefone: true
+        }
+    })
+    res.json(peopleWithTelefone)
 }
 
-export const deletePeople = async(req: Request, res: Response) => {
+export const deletePeople = async (req: Request, res: Response) => {
     const { id } = req.params
-    await prisma.pessoa.delete({
-        where: { id: Number(id) }
-    })
-    res.status(204).send()
+    try {
+        const pessoa = await prisma.pessoa.findUnique({
+            where: {
+                id: Number(id)
+            }
+        })
+
+        if (!pessoa) {
+            return res.status(404).json({ error: "Pessoa não encontrada" });
+        }
+
+        await prisma.telefone.deleteMany({
+            where: { pessoaId: Number(id) }
+        })
+        await prisma.pessoa.delete({
+            where: { id: Number(id) }
+        })
+        res.status(204).send()
+    } catch (e) {
+        console.error("Erro ao deletar pessoa:", e);
+        res.status(500).json({ error: "Erro ao deletar pessoa" });
+    }
 }
